modified structure adding two more agents
